- pull up force2D to SelectiveEvent and simplify GetFeatures#buildGeoFragment() - move some functionality related to GetFeaturesByGeometryEvent handling into GetFeaturesByGeometry QR
