+++ xyz-models/src/main/java/com/here/xyz/events/GetFeaturesByIdEvent.java
@@ -29,7 +29,6 @@ import java.util.List;
 public final class GetFeaturesByIdEvent extends SelectiveEvent<GetFeaturesByIdEvent> {
 
   private List<String> ids;
-  private boolean force2D;
 
   public List<String> getIds() {
     return this.ids;
@@ -41,21 +40,6 @@ public final class GetFeaturesByIdEvent extends SelectiveEvent<GetFeaturesByIdEv
 
   public GetFeaturesByIdEvent withIds(List<String> ids) {
     setIds(ids);
-    return this;
-  }
-
-  public boolean isForce2D() {
-    return force2D;
-  }
-
-  @SuppressWarnings("WeakerAccess")
-  public void setForce2D(boolean force2D) {
-    this.force2D = force2D;
-  }
-
-  @SuppressWarnings("unused")
-  public GetFeaturesByIdEvent withForce2D(boolean force2D) {
-    setForce2D(force2D);
     return this;
   }
 }
