+++ xyz-models/src/main/java/com/here/xyz/events/SearchForFeaturesEvent.java
@@ -30,7 +30,6 @@ public class SearchForFeaturesEvent<T extends SearchForFeaturesEvent> extends Qu
   private static final long MAX_LIMIT = 100_000L;
 
   private long limit = DEFAULT_LIMIT;
-  private boolean force2D;
 
   public long getLimit() {
     return limit;
@@ -44,23 +43,6 @@ public class SearchForFeaturesEvent<T extends SearchForFeaturesEvent> extends Qu
   @SuppressWarnings("unused")
   public T withLimit(long limit) {
     setLimit(limit);
-    //noinspection unchecked
-    return (T) this;
-  }
-
-  @SuppressWarnings("WeakerAccess")
-  public boolean isForce2D() {
-    return force2D;
-  }
-
-  @SuppressWarnings("WeakerAccess")
-  public void setForce2D(boolean force2D) {
-    this.force2D = force2D;
-  }
-
-  @SuppressWarnings("unused")
-  public T withForce2D(boolean force2D) {
-    setForce2D(force2D);
     //noinspection unchecked
     return (T) this;
   }
