+++ xyz-models/src/main/java/com/here/xyz/events/SelectiveEvent.java
@@ -21,9 +21,10 @@ package com.here.xyz.events;
 
 import java.util.List;
 
-public class SelectiveEvent<T extends ContextAwareEvent> extends ContextAwareEvent<T> {
+public class SelectiveEvent<T extends SelectiveEvent> extends ContextAwareEvent<T> {
 
   private List<String> selection;
+  private boolean force2D;
 
   @SuppressWarnings("unused")
   public List<String> getSelection() {
@@ -39,5 +40,22 @@ public class SelectiveEvent<T extends ContextAwareEvent> extends ContextAwareEve
   public T withSelection(List<String> selection) {
     setSelection(selection);
     return (T)this;
+  }
+
+  @SuppressWarnings("WeakerAccess")
+  public boolean isForce2D() {
+    return force2D;
+  }
+
+  @SuppressWarnings("WeakerAccess")
+  public void setForce2D(boolean force2D) {
+    this.force2D = force2D;
+  }
+
+  @SuppressWarnings("unused")
+  public T withForce2D(boolean force2D) {
+    setForce2D(force2D);
+    //noinspection unchecked
+    return (T) this;
   }
 }
