+++ xyz-psql-connector/src/main/java/com/here/xyz/psql/SQLQueryBuilder.java
@@ -22,7 +22,6 @@ import static com.here.xyz.events.GetFeaturesByTileEvent.ResponseType.GEO_JSON;
 
 import com.here.xyz.events.Event;
 import com.here.xyz.events.GetFeaturesByBBoxEvent;
-import com.here.xyz.events.GetFeaturesByGeometryEvent;
 import com.here.xyz.events.GetFeaturesByTileEvent;
 import com.here.xyz.events.GetFeaturesByTileEvent.ResponseType;
 import com.here.xyz.events.GetHistoryStatisticsEvent;
@@ -34,12 +33,11 @@ import com.here.xyz.events.SpatialQueryEvent;
 import com.here.xyz.models.geojson.HQuad;
 import com.here.xyz.models.geojson.WebMercatorTile;
 import com.here.xyz.models.geojson.coordinates.BBox;
-import com.here.xyz.models.geojson.coordinates.WKTHelper;
-import com.here.xyz.models.geojson.implementation.Geometry;
 import com.here.xyz.psql.config.PSQLConfig;
 import com.here.xyz.psql.factory.H3SQL;
 import com.here.xyz.psql.factory.QuadbinSQL;
 import com.here.xyz.psql.factory.TweaksSQL;
+import com.here.xyz.psql.query.GetFeaturesByGeometry;
 import com.here.xyz.psql.query.ModifySpace;
 import com.here.xyz.psql.query.SearchForFeatures;
 import com.here.xyz.psql.tools.DhString;
@@ -69,27 +67,7 @@ public class SQLQueryBuilder {
         return new SQLQuery("SELECT nextval('${schema}.\"" + table.replaceAll("-","_") + "_hst_seq\"')");
     }
 
-    public static SQLQuery buildGetFeaturesByGeometryQuery(GetFeaturesByGeometryEvent event)
-        throws SQLException{
-        final int radius = event.getRadius();
-        final Geometry geometry = event.getGeometry();
-
-        final SQLQuery geoQuery;
-
-        if(event.getH3Index() != null){
-            if(radius != 0)
-                geoQuery = new SQLQuery("ST_Intersects( geo, ST_Buffer(hexbin::geography,? )::geometry)", radius);
-            else
-                geoQuery = new SQLQuery("ST_Intersects( geo, hexbin)");
-        }else{
-            geoQuery = radius != 0 ? new SQLQuery("ST_Intersects(geo, ST_Buffer(ST_GeomFromText('"
-                    + WKTHelper.geometryToWKB(geometry) + "')::geography, ? )::geometry)", radius) : new SQLQuery("ST_Intersects(geo, ST_GeomFromText('"
-                    + WKTHelper.geometryToWKB(geometry) + "',4326))");
-        }
-        return generateCombinedQuery(event, geoQuery);
-    }
-
-    public static SQLQuery buildGetFeaturesByBBoxQuery(final GetFeaturesByBBoxEvent event)
+  public static SQLQuery buildGetFeaturesByBBoxQuery(final GetFeaturesByBBoxEvent event)
         throws SQLException{
         final BBox bbox = event.getBbox();
 
@@ -739,82 +717,10 @@ public class SQLQueryBuilder {
     private static SQLQuery generateCombinedQueryTweaks(GetFeaturesByBBoxEvent event, SQLQuery indexedQuery, String tweaksgeo, boolean bTestTweaksGeoIfNull)
     { return generateCombinedQueryTweaks(event, indexedQuery, tweaksgeo, bTestTweaksGeoIfNull, -1.0f, false );  }
 
-    private static SQLQuery generateCombinedQuery(SpatialQueryEvent event, SQLQuery indexedQuery)
-    {
-      SQLQuery searchQuery = generateSearchQuery(event);
-      boolean bConvertGeo2Geojson = event instanceof GetFeaturesByGeometryEvent || getResponseType((GetFeaturesByBBoxEvent) event) == GEO_JSON;
+  private static SQLQuery generateCombinedQuery(SpatialQueryEvent event, SQLQuery indexedQuery) {
+      return GetFeaturesByGeometry.generateCombinedQueryBWC(event, indexedQuery);
+  }
 
-        final SQLQuery query = new SQLQuery();
-
-        String h3Index = event instanceof GetFeaturesByGeometryEvent ? ((GetFeaturesByGeometryEvent) event).getH3Index() : null;
-
-        if(h3Index != null)
-            query.append("WITH h AS (SELECT h3ToGeoBoundaryDeg( ('x' || '"+h3Index+"' )::bit(60)::bigint ) as hexbin )");
-
-        query.append("SELECT ");
-
-        query.append(SQLQuery.selectJson(event));
-        query.append(",");
-        query.append(geometrySelectorForEvent( event, bConvertGeo2Geojson));
-
-        if(h3Index != null)
-            query.append("FROM ${schema}.${table} ${{tableSample}}, h WHERE");
-        else
-            query.append("FROM ${schema}.${table} ${{tableSample}} WHERE");
-        query.setQueryFragment("tableSample", ""); //Can be overridden by caller
-
-        query.append(indexedQuery);
-
-        if( searchQuery != null )
-        { query.append(" and ");
-            query.append(searchQuery);
-        }
-
-        query.append(" ${{orderBy}} ");
-        query.setQueryFragment("orderBy", ""); //Can be overridden by caller
-
-        query.append("LIMIT ?", event.getLimit());
-        return query;
-    }
-
-    /**
-     * Returns the query, which will contains the geometry object.
-     */
-    private static SQLQuery geometrySelectorForEvent(final SpatialQueryEvent event, boolean bGeoJson) {
-        String forceMode = getForceMode(event.isForce2D());
-
-        if(!event.getClip()){
-            return (bGeoJson ?
-                    new SQLQuery("replace(ST_AsGeojson("+forceMode+"(geo),?::INTEGER),'nan','0') as geo", GEOMETRY_DECIMAL_DIGITS)
-                    : new SQLQuery(forceMode+"(geo) as geo"));
-        }else{
-            if(event instanceof GetFeaturesByBBoxEvent){
-                final BBox bbox = ((GetFeaturesByBBoxEvent)event).getBbox();
-                String geoSqlAttrib = (bGeoJson ? DhString.format("replace(ST_AsGeoJson(ST_Intersection(ST_MakeValid(geo),ST_MakeEnvelope(?,?,?,?,4326)),%d),'nan','0') as geo", GEOMETRY_DECIMAL_DIGITS)
-                        : "ST_Intersection( ST_MakeValid(geo),ST_MakeEnvelope(?,?,?,?,4326) ) as geo");
-
-                return new SQLQuery(geoSqlAttrib, bbox.minLon(), bbox.minLat(), bbox.maxLon(), bbox.maxLat());
-            }else if(event instanceof GetFeaturesByGeometryEvent){
-                //Clip=true =>  Use input Geometry for clipping
-                final Geometry geometry = ((GetFeaturesByGeometryEvent)event).getGeometry();
-                //If h3Index is given - use it as input geometry
-                final String h3Index =  ((GetFeaturesByGeometryEvent)event).getH3Index();
-
-                String wktGeom = h3Index == null ? "ST_GeomFromText('" + WKTHelper.geometryToWKB(geometry) + "',4326)" : "hexbin" ;
-
-                //If radius is not null
-                if(((GetFeaturesByGeometryEvent)event).getRadius() != 0){
-                    //Enlarge input geometry with ST_Buffer
-                    wktGeom = (h3Index == null ?
-                            DhString.format("ST_Buffer(ST_GeomFromText('" + WKTHelper.geometryToWKB(geometry) + "')::geography, %d )::geometry", ((GetFeaturesByGeometryEvent)event).getRadius())
-                            : DhString.format("ST_Buffer(hexbin::geography,%d)::geometry", ((GetFeaturesByGeometryEvent)event).getRadius()));
-                }
-                return new SQLQuery("replace(ST_AsGeoJson(ST_Intersection(ST_MakeValid(geo),"+wktGeom+"),?::INTEGER),'nan','0') as geo", GEOMETRY_DECIMAL_DIGITS);
-            }
-        }
-        //Should not happen (currently only used with GetFeaturesByBBoxEvent / GetFeaturesByGeometryEvent)
-        return new SQLQuery("ST_AsGeojson(geo) as geo");
-    }
 
   protected static SQLQuery generateSearchQuery(final QueryEvent event) {
       return SearchForFeatures.generateSearchQueryBWC(event);
