+++ xyz-psql-connector/src/main/java/com/here/xyz/psql/query/GetFeatures.java
@@ -154,12 +154,8 @@ public abstract class GetFeatures<E extends ContextAwareEvent> extends ExtendedS
   }
 
   protected String buildGeoFragment(E event) {
-    boolean isForce2D = false;
-    if (event instanceof GetFeaturesByIdEvent)
-      isForce2D = ((GetFeaturesByIdEvent) event).isForce2D();
-    else if (event instanceof SearchForFeaturesEvent)
-      isForce2D = ((SearchForFeaturesEvent<?>) event).isForce2D();
-
-    return "replace(ST_AsGeojson(" + (isForce2D ? "ST_Force2D" : "ST_Force3D") + "(geo), " + SQLQueryBuilder.GEOMETRY_DECIMAL_DIGITS + "), 'nan', '0')";
+    boolean isForce2D = event instanceof SelectiveEvent ? ((SelectiveEvent) event).isForce2D() : false;
+    return "replace(ST_AsGeojson(" + (isForce2D ? "ST_Force2D" : "ST_Force3D") + "(geo), "
+        + SQLQueryBuilder.GEOMETRY_DECIMAL_DIGITS + "), 'nan', '0')";
   }
 }
